requirements.txt
Pillow==4.1.0
PyYaml==3.12
